Added cloud MongoDB connection and data storing, more...
Created a cloud MongoDB cluster and made loop store all data as documents in the regional_weather collection.

Take into account authentication is required, so you need an user account (ask me). Once you have one, add it to environment variables. Installation of MongoDB on the OS and PyCharm driver (if applies) is also required. MongoDB Compass is an amazing tool for visualizing the cluster as well as work with it (great interface compared to raw shell commands)

Todo: Refine data storing in DB (bucket patterns).