Added bucket pattern, improve index performance and data access...
All data from the same day is saved in the same document by updating the values and saving all history in an array. This make it easier to access while saving a lot of space.

Reduced redundancy by adding checks to avoid storing the same data.

Added and improved comments.

Todo: Find a way to flag the 'Waiting for data...' message so it only shows once per wait (no console spam).